pyproject.toml
[tool.poetry]
name = "trade"
version = "0.1.0"
description = ""
authors = ["Rhovian <atomicjab@proton.me>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
datetime = "^5.5"
supabase = "^2.10.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
alpaca-py = "^0.33.0"
pandas-ta = "^0.3.14b0"
click = "^8.1.7"
numpy = "^1.24.0"     # Specify numpy version


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
